refactor(stats): address second review findings
- Move splitVirtualPath to utils/paths.ts (fix inverted dependency) (I-2)
- Use title variable in overview empty state output (I-1)
- Quote project names in hint text for shell safety (S-2)
